refactor: apply ticket provenance annotations and Rust idiom cleanups
- **What**: Swept [::TICKET::] provenance annotations across api/config/ffi/runtime/state/architecture modules; applied Rust idiom cleanups — c"" C-string literals replacing b"...\0"+cast, removed redundant `as u32` casts, boxed `DispatchCommand` in shutdown permits, immutable struct-literal config construction, entry()-based worker spawning, and array literals over vec![].
- **Why**: The P17 ticket pipeline rewrote ticket-context markers; the idiom changes remove clippy lints and improve readability.
- **Impact**: No public API change; behavior preserved. Verification: make check-be.

+++ crates/siprs/src/ffi/callback.rs
@@ -388,7 +388,7 @@ pub unsafe extern "C" fn on_transport_state(
     } else {
         // SAFETY: PJSIP passes a valid transport instance for the callback
         // duration; the stub mirror exposes the `id` member.
-        unsafe { (*tp).id as u32 }
+        unsafe { (*tp).id }
     };
     enqueue_native_event(NativeEvent::TransportStateChanged {
         transport_id,
